requirements.txt
python-dotenv>=1.0.0
rich>=13.7.0
openai>=1.3.0
aiohttp>=3.9.1
asyncio>=3.4.3
tiktoken>=0.5.1
openai-agents>=0.0.9
pyperclip>=1.8.2
questionary>=2.0.0